use src folder for metro and build dir for node in metro-runtime
