+++ packages/@expo/metro-runtime/package.json
@@ -2,7 +2,10 @@
   "name": "@expo/metro-runtime",
   "version": "3.0.1",
   "description": "Tools for making advanced Metro bundler features work",
+  "types": "build",
   "main": "build",
+  "browser": "src",
+  "react-native": "src",
   "homepage": "https://github.com/expo/expo/tree/main/packages/@expo/metro-runtime",
   "keywords": [],
   "author": "650 Industries, Inc.",
