+++ packages/expo-router/src/views/ErrorBoundary.tsx
@@ -54,10 +54,10 @@ if (process.env.NODE_ENV === 'development') {
 let StackTrace: React.FC<{ logData: LogBoxLog | null }>;
 
 if (process.env.NODE_ENV === 'development') {
-  const { LogContext } = require('@expo/metro-runtime/build/error-overlay/Data/LogContext');
+  const { LogContext } = require('@expo/metro-runtime/src/error-overlay/Data/LogContext');
   const {
     LogBoxInspectorStackFrames,
-  } = require('@expo/metro-runtime/build/error-overlay/overlay/LogBoxInspectorStackFrames');
+  } = require('@expo/metro-runtime/src/error-overlay/overlay/LogBoxInspectorStackFrames');
 
   StackTrace = function ({ logData }: { logData: LogBoxLog | null }) {
     if (!logData?.symbolicated?.stack?.stack) {
